review: hold the authority through the write, and leave the merge arm alone
Codex found two High defects in the first shape of this fix, one of which
Fable had also raised.

The probe sat before the disposition switch, so it ran on the merge arm too.
That is not the no-op the last commit claimed. mergePair answers a BARE
CONFLICT for a target the caller cannot change — deliberately, so a refusal
names no record the caller has not already been handed — and an outer probe
pre-empted it with 403. It also moved authority ahead of winner_id validation,
turning a 422 into a 403. The probe now sits in the not_a_duplicate arm, merge
keeps its own refusal, and TestDedupeMergeArmKeepsItsOwnRefusal pins it: with
the old placement that test reports permission denied where it wants conflict.

The probe also committed its own transaction two transactions before the write,
so a grant revoked in between left the write standing on authority that was
already gone. writePairDecision now runs the probe and the write in ONE
transaction. The write is the parameter and the probe is not: a caller can
choose which decision to record, never whether the guard runs. The paths that
must not probe — disposeMerge's mark and its compensating rollback — do not
come through it at all, so nothing is decided by a flag.

Three smaller ones. The helper comment claimed every seat holding person:update
passes requireDedupeWrite over every colleague's records; that guard asks for
update on the pair's own type, so the claim was true of one type and stated of
three. The 'must be 403' assertion only rejected ErrNotFound, which nil and a
conflict both pass. And the half-writable test covered dismiss on person alone,
so an undo checking one endpoint, or a helper wired to person, would have
passed — it now runs both verbs across both sides and both entity types.

Two consequences of the refactor. The file crossed the 500-line cap, so the
decision verbs move to dedupedisposition.go and dedupequeue.go keeps the
reading half. And the audit calls moved into the new Tx cores, so the two
write-shape waivers are re-keyed to the functions that now hold them; their
reasons are unchanged, and they are still about events rather than authority.

Refs #1949
