+++ backend/internal/modules/people/dedupequeue.go
@@ -23,7 +23,6 @@ import (
 	"github.com/jackc/pgx/v5"
 
 	"github.com/gradionhq/margince/backend/internal/platform/auth"
-	"github.com/gradionhq/margince/backend/internal/platform/database/storekit"
 	"github.com/gradionhq/margince/backend/internal/shared/apperrors"
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/ids"
 	"github.com/gradionhq/margince/backend/internal/shared/kernel/principal"
@@ -284,189 +283,4 @@ func readDedupeCandidate(ctx context.Context, tx pgx.Tx, id ids.UUID) (DedupeCan
 		return r, fmt.Errorf("people: reading dedupe candidate: %w", err)
 	}
 	return r, nil
-}
-
-// ensurePairWritable narrows the pair's read gate to the authority a decision
-// needs. A disposition CHANGES both records the candidate names — a dismissal
-// suppresses them as duplicates for the whole workspace, an undo puts the pair
-// back into everyone's queue — so each end carries the write-authority probe
-// rather than the visibility one, exactly as mergePair states it for the merge
-// verb: a colleague handed a `read` share of either record may not spend it on
-// the queue's verdict.
-//
-// The object grant is not that authority and never was. Person, organization
-// and lead are workspace-readable identity, so every seat holding
-// person:update passes requireDedupeWrite over every colleague's records.
-//
-// It refuses with 403, not 404: GetDedupeCandidate has already told this caller
-// the pair is theirs to read, so there is nothing left for existence-hiding to
-// hide (platform/auth/writescope.go states the same rule for every exported
-// spelling). row.EntityType is the table name, and all three are row-scoped.
-func (s *Store) ensurePairWritable(ctx context.Context, entityType string, left, right ids.UUID) error {
-	return s.db.Tx(ctx, func(tx pgx.Tx) error {
-		if err := auth.EnsureWritable(ctx, tx, entityType, left); err != nil {
-			return err
-		}
-		return auth.EnsureWritable(ctx, tx, entityType, right)
-	})
-}
-
-// DisposeDedupeCandidate decides one pair. merge executes the owner's
-// merge verb with the LOSER folding into the winner; not_a_duplicate
-// suppresses the pair forever. Human-only (the transport enforces the
-// x-agent-access posture; the store re-checks the principal).
-func (s *Store) DisposeDedupeCandidate(ctx context.Context, id ids.UUID, disposition string, winnerID *ids.UUID) (DedupeCandidateRow, error) {
-	actor, ok := principal.Actor(ctx)
-	if !ok || actor.Type != principal.PrincipalHuman {
-		return DedupeCandidateRow{}, fmt.Errorf("people: only a human disposes a dedupe pair: %w", apperrors.ErrPermissionDenied)
-	}
-	row, err := s.GetDedupeCandidate(ctx, id)
-	if err != nil {
-		return DedupeCandidateRow{}, err
-	}
-	if err := requireDedupeWrite(ctx, row.EntityType); err != nil {
-		return DedupeCandidateRow{}, err
-	}
-	// One site, both arms. The probe sits HERE rather than inside
-	// setDedupeDisposition, which serves the not_a_duplicate arm and
-	// disposeMerge alike; the merge arm then holds it twice — once here and
-	// once through mergePair — which is harmless, and the alternative is a
-	// verb that decides the queue without it.
-	if err := s.ensurePairWritable(ctx, row.EntityType, row.LeftID, row.RightID); err != nil {
-		return DedupeCandidateRow{}, err
-	}
-	if row.Disposition != dispositionOpen {
-		return DedupeCandidateRow{}, fmt.Errorf("people: candidate already disposed (%s): %w", row.Disposition, apperrors.ErrConflict)
-	}
-
-	switch disposition {
-	case dispositionNotDuplicate:
-		if err := s.setDedupeDisposition(ctx, id, dispositionNotDuplicate, actor.UserID); err != nil {
-			return DedupeCandidateRow{}, err
-		}
-	case "merge":
-		if err := s.disposeMerge(ctx, id, row, winnerID, actor.UserID); err != nil {
-			return DedupeCandidateRow{}, err
-		}
-	default:
-		return DedupeCandidateRow{}, &DedupeInputError{Field: "disposition", Msg: "must be merge or not_a_duplicate"}
-	}
-	return s.GetDedupeCandidate(ctx, id)
-}
-
-// disposeMerge is the merge arm: validate the winner, mark first (a CAS on
-// open, so a concurrent decision cannot double-merge), then run the ONE
-// merge verb in its own transaction. A merge failure re-opens the row —
-// the queue never claims a merge that did not happen.
-func (s *Store) disposeMerge(ctx context.Context, id ids.UUID, row DedupeCandidateRow, winnerID *ids.UUID, by ids.UUID) error {
-	if winnerID == nil || (*winnerID != row.LeftID && *winnerID != row.RightID) {
-		return &DedupeInputError{Field: "winner_id", Msg: "must be one of the pair"}
-	}
-	loser := row.LeftID
-	if loser == *winnerID {
-		loser = row.RightID
-	}
-	if err := s.setDedupeDisposition(ctx, id, dispositionMerged, by); err != nil {
-		return err
-	}
-	if err := s.executeDedupeMerge(ctx, row.EntityType, loser, *winnerID); err != nil {
-		if reopenErr := s.reopenDedupeCandidate(ctx, id); reopenErr != nil {
-			return errors.Join(err, reopenErr)
-		}
-		return err
-	}
-	return nil
-}
-
-// executeDedupeMerge runs the ONE merge implementation for the pair's type.
-func (s *Store) executeDedupeMerge(ctx context.Context, entityType string, loser, winner ids.UUID) error {
-	switch entityType {
-	case entityPerson:
-		_, err := s.MergePerson(ctx, ids.From[ids.PersonKind](loser), ids.From[ids.PersonKind](winner))
-		return err
-	case entityOrganization:
-		_, err := s.MergeOrganization(ctx, ids.From[ids.OrganizationKind](loser), ids.From[ids.OrganizationKind](winner))
-		return err
-	case entityLead:
-		_, err := s.MergeLead(ctx, ids.From[ids.LeadKind](loser), ids.From[ids.LeadKind](winner))
-		return err
-	default:
-		return fmt.Errorf("people: unmergeable entity type %q", entityType)
-	}
-}
-
-// setDedupeDisposition is the CAS open→disposed; losing the race answers
-// conflict, never a second merge. The audit row rides the same commit;
-// dedupe_candidate is not a §4.1 stream entity, so the disposition has no
-// bus event — the audit ledger is the record (the merge arm's
-// person.merged/organization.merged carries the bus-visible fact).
-func (s *Store) setDedupeDisposition(ctx context.Context, id ids.UUID, disposition string, by ids.UUID) error {
-	return s.db.Tx(ctx, func(tx pgx.Tx) error {
-		tag, err := tx.Exec(ctx, `
-			UPDATE dedupe_candidate SET disposition = $2, disposed_by = $3, disposed_at = now()
-			WHERE id = $1 AND disposition = 'open'`, id, disposition, by)
-		if err != nil {
-			return fmt.Errorf("people: disposing dedupe candidate: %w", err)
-		}
-		if tag.RowsAffected() == 0 {
-			return fmt.Errorf("people: candidate already disposed: %w", apperrors.ErrConflict)
-		}
-		_, err = storekit.Audit(ctx, tx, "resolve", auditEntityDedupe, id,
-			map[string]any{auditKeyDisposition: dispositionOpen},
-			map[string]any{auditKeyDisposition: disposition})
-		return err
-	})
-}
-
-func (s *Store) reopenDedupeCandidate(ctx context.Context, id ids.UUID) error {
-	return s.db.Tx(ctx, func(tx pgx.Tx) error {
-		tag, err := tx.Exec(ctx, `
-			UPDATE dedupe_candidate SET disposition = 'open', disposed_by = NULL, disposed_at = NULL
-			WHERE id = $1 AND disposition <> 'open'`, id)
-		if err != nil {
-			return err
-		}
-		if tag.RowsAffected() == 0 {
-			// Already open (a concurrent undo) — the desired state holds.
-			return nil
-		}
-		_, err = storekit.Audit(ctx, tx, "restore", auditEntityDedupe, id,
-			nil, map[string]any{auditKeyDisposition: dispositionOpen})
-		return err
-	})
-}
-
-// UndoDedupeDisposition re-opens a dismissed pair (the suppression lifts).
-// A merged pair answers ErrNotUndoable: reversing a merge needs the merge
-// verb's own reversibility (PO-AC-M6), which does not exist yet — the
-// queue must not pretend otherwise.
-func (s *Store) UndoDedupeDisposition(ctx context.Context, id ids.UUID) (DedupeCandidateRow, error) {
-	actor, ok := principal.Actor(ctx)
-	if !ok || actor.Type != principal.PrincipalHuman {
-		return DedupeCandidateRow{}, fmt.Errorf("people: only a human re-opens a dedupe pair: %w", apperrors.ErrPermissionDenied)
-	}
-	row, err := s.GetDedupeCandidate(ctx, id)
-	if err != nil {
-		return DedupeCandidateRow{}, err
-	}
-	if err := requireDedupeWrite(ctx, row.EntityType); err != nil {
-		return DedupeCandidateRow{}, err
-	}
-	// Probed HERE and not inside reopenDedupeCandidate, which has a second
-	// caller that is not a user act: disposeMerge reopens the row as the
-	// compensating rollback of a merge that failed. A grant revoked mid-flight
-	// would then strand a candidate at 'merged' with no merge behind it.
-	if err := s.ensurePairWritable(ctx, row.EntityType, row.LeftID, row.RightID); err != nil {
-		return DedupeCandidateRow{}, err
-	}
-	switch row.Disposition {
-	case dispositionOpen:
-		return DedupeCandidateRow{}, fmt.Errorf("people: candidate is already open: %w", apperrors.ErrConflict)
-	case dispositionMerged:
-		return DedupeCandidateRow{}, fmt.Errorf("%w: %w", ErrNotUndoable, apperrors.ErrConflict)
-	}
-	if err := s.reopenDedupeCandidate(ctx, id); err != nil {
-		return DedupeCandidateRow{}, err
-	}
-	return s.GetDedupeCandidate(ctx, id)
 }
