+++ backend/writeshape_test.go
@@ -105,8 +105,8 @@ var auditOnlyWrites = gatekit.Waive(map[string]string{
 	"internal/modules/deals:UpdateOfferTemplate":            "offer templates are workspace-authored config, the Product/Quota precedent \u2014 the closed catalog (events.md \u00a75) defines no offer_template.* type and the closed-verb law forbids inventing one build-side",
 	"internal/modules/deals:ArchiveOfferTemplate":           "offer templates are workspace-authored config, the Product/Quota precedent \u2014 the closed catalog (events.md \u00a75) defines no offer_template.* type and the closed-verb law forbids inventing one build-side",
 	"internal/modules/people:createOrJoinSiteRead":          "the deep-read dossier is operational crawl status, not a record fact \u2014 the closed catalog (events.md \u00a75) defines no site_read.* type; the facts a read produces land through its staged proposals, each emitting its own event on accept",
-	"internal/modules/people:setDedupeDisposition":          "the queue verdict is review-flow state, not a record fact \u2014 the closed catalog (events.md \u00a75) defines no dedupe.* type; the merge arm's person.merged/organization.merged carries the bus-visible outcome",
-	"internal/modules/people:reopenDedupeCandidate":         "the queue verdict is review-flow state, not a record fact \u2014 the closed catalog (events.md \u00a75) defines no dedupe.* type; the merge arm's person.merged/organization.merged carries the bus-visible outcome",
+	"internal/modules/people:setDedupeDispositionTx":        "the queue verdict is review-flow state, not a record fact \u2014 the closed catalog (events.md \u00a75) defines no dedupe.* type; the merge arm's person.merged/organization.merged carries the bus-visible outcome",
+	"internal/modules/people:reopenDedupeCandidateTx":       "the queue verdict is review-flow state, not a record fact \u2014 the closed catalog (events.md \u00a75) defines no dedupe.* type; the merge arm's person.merged/organization.merged carries the bus-visible outcome",
 	"internal/modules/identity:CreateRecordGrant":           "the closed catalog (events.md \u00a75) defines no grant.* type; the closed-verb law forbids inventing one build-side",
 	"internal/modules/identity:RevokeRecordGrant":           "the closed catalog (events.md \u00a75) defines no grant.* type; the closed-verb law forbids inventing one build-side",
 	"internal/modules/identity:mintPassport":                "minting an Agent Seat Passport is ratified audit-only \u2014 the closed catalog (events.md \u00a75.6a) defines passport.revoked and NO issuance counterpart, so the only bus-visible passport fact is its death. That asymmetry is deliberate on the consumer side (a long-lived holder must drop a credential that died; nobody has to react to one being born) and the closed-verb law forbids inventing passport.issued build-side. The mint is attributable via its audit row; the missing type is raised upstream (P3)",
